* Converted sources to use standard C9x int types defined in inttypes.h.   Corrected a number of format specifiers for printf/scanf.  Changed   a couple variable types here and there to correct possible problems.
  This provides a good way to print 64-bit ints with either standard GNU
  C libraries or the MSVCRT printf under Windows.  Any future code should
  print s32s as PRId32, u64s as PRIu64, and so on.  The corresponding
  scanf format specifiers are SCNd32/SCNx32 and SCNu64.  Variables declared
  without bitlengths should still use non-macro formats (%d for int, %ld for
  long).  At the moment, size_t is printed as %lu, unless someone has a
  better suggestion as to what it should be.


git-svn-id: https://ggnfs.svn.sourceforge.net/svnroot/ggnfs/trunk@29 80ef69ca-9341-0410-9d60-8acc7774fa31

+++ src/blanczos64.c
@@ -682,13 +682,13 @@ int blockLanczos64(u64 *deps, MAT_MULT_FUNC_PTR64 MultB,
               (double)100.0*64.0*iterations/(1.02*n), now-startTime, estTotal);  
     if ((double)100.0*64.0*iterations/n > 250) {
       fprintf(stderr, "Some error has occurred: Lanczos is not converging!\n");
-      fprintf(stderr, "Number of iterations is %ld.\n", iterations);
+      fprintf(stderr, "Number of iterations is %" PRIu32 ".\n", iterations);
       /* Add some debugging stuff here! */
       fprintf(stderr, "Terminating...\n");
       exit(-1);
     }
   } while (cont);
-  printf("\nBlock Lanczos used %ld iterations.\n", iterations);
+  printf("\nBlock Lanczos used %" PRIu32 " iterations.\n", iterations);
 
           
   if (malloc_aligned64(Z, 16, 2*n)) Z=NULL;
@@ -806,7 +806,7 @@ int blockLanczos64(u64 *deps, MAT_MULT_FUNC_PTR64 MultB,
     if (tmp_n[i])
       isZero = 0;
   if (!(isZero))
-    printf("Some error occurred: Final product (B)(deps) is nonzero (i=%ld)!\n", (s32)i);
+    printf("Some error occurred: Final product (B)(deps) is nonzero (i=%" PRIu64 ")!\n", i);
   else
     printf("Verified.\n");
